avoiding unsigned underflow. refs #1316
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@16499 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/utils/geom/PositionVector.cpp
@@ -191,7 +191,7 @@ PositionVector::operator[](int index) const {
     if (index >= 0) {
         return at(index);
     } else {
-        return at(size() + index);
+        return at((int)size() + index);
     }
 }
 
@@ -201,7 +201,7 @@ PositionVector::operator[](int index) {
     if (index >= 0) {
         return at(index);
     } else {
-        return at(size() + index);
+        return at((int)size() + index);
     }
 }
